Refactor permission flow and remove test image
