+++ README.md
@@ -41,7 +41,7 @@ npm run dev:https
 
 - **권한 거부**: 브라우저 설정에서 카메라 권한 확인
 - **카메라 사용 중**: 다른 애플리케이션에서 카메라 사용 종료
-- **카메라 없음**: 자동으로 테스트 이미지 모드로 전환
+- **카메라 없음**: 카메라 연결 상태를 확인하세요
 
 ### React Webcam 사용 예시
