+++ src/constants/index.ts
@@ -36,21 +36,20 @@ export const CLAUDE_CONFIG = {
 // 에러 메시지
 export const ERROR_MESSAGES = {
   camera: {
-    notFound: "카메라를 찾을 수 없습니다. 테스트 이미지를 사용합니다.",
-    notAllowed: "카메라 권한이 거부되었습니다. 테스트 이미지를 사용합니다.",
-    unknown: "카메라에 접근할 수 없습니다. 테스트 이미지를 사용합니다.",
-    inUse: "카메라가 다른 애플리케이션에서 사용 중입니다. 테스트 이미지를 사용합니다.",
-    constraintError: "요청한 카메라 설정을 지원하지 않습니다. 테스트 이미지를 사용합니다.",
-    notSupported: "브라우저에서 카메라를 지원하지 않습니다. 테스트 이미지를 사용합니다.",
-    aborted: "카메라 접근이 중단되었습니다. 테스트 이미지를 사용합니다.",
-    httpsRequired: "HTTPS 환경에서만 카메라를 사용할 수 있습니다. 테스트 이미지를 사용합니다.",
-    apiNotSupported: "브라우저에서 카메라 API를 지원하지 않습니다. 테스트 이미지를 사용합니다.",
+    notFound: "카메라를 찾을 수 없습니다.",
+    notAllowed: "카메라 권한이 거부되었습니다.",
+    unknown: "카메라에 접근할 수 없습니다.",
+    inUse: "카메라가 다른 애플리케이션에서 사용 중입니다.",
+    constraintError: "요청한 카메라 설정을 지원하지 않습니다.",
+    notSupported: "브라우저에서 카메라를 지원하지 않습니다.",
+    aborted: "카메라 접근이 중단되었습니다.",
+    httpsRequired: "HTTPS 환경에서만 카메라를 사용할 수 있습니다.",
+    apiNotSupported: "브라우저에서 카메라 API를 지원하지 않습니다.",
   },
   image: {
     invalidBase64: "Invalid base64 data",
     tooSmall: "Image data too small",
     captureFailure: "이미지 캡처에 실패했습니다. 다시 시도해주세요.",
-    testImageLoadFailure: "테스트 이미지를 로드할 수 없습니다.",
   },
   api: {
     imageRequired: "Image is required",
