+++ src/hooks/useCamera.ts
@@ -10,9 +10,8 @@ export function useCamera() {
   const webcamRef = useRef<Webcam>(null);
   const [isReady, setIsReady] = useState(false);
   const [error, setError] = useState<CameraError | null>(null);
-  const [useTestImage, setUseTestImage] = useState(false);
   const [isRetrying, setIsRetrying] = useState(false);
-  const [permissionRequested, setPermissionRequested] = useState(false);
+  const [permissionRequested, setPermissionRequested] = useState(true);
 
   const videoConstraints: MediaTrackConstraints = {
     width: { ideal: CAMERA_CONFIG.idealWidth },
@@ -27,41 +26,38 @@ export function useCamera() {
     }
     setIsReady(false);
     setError(null);
-    setUseTestImage(false);
   }, []);
 
   const handleUserMedia = useCallback(() => {
     setIsReady(true);
     setError(null);
-    setUseTestImage(false);
   }, []);
 
   const handleUserMediaError = useCallback((err: string | DOMException) => {
     console.error("Camera access failed:", err);
     let errorType: CameraErrorType = "unknown";
-    let message = "카메라에 접근할 수 없습니다. 테스트 이미지를 사용합니다.";
+    let message = "카메라에 접근할 수 없습니다.";
 
     if (err instanceof DOMException) {
       if (err.name === "NotAllowedError" || err.name === "PermissionDeniedError") {
         errorType = "not-allowed";
-        message = "카메라 권한이 거부되었습니다. 테스트 이미지를 사용합니다.";
+        message = "카메라 권한이 거부되었습니다.";
       } else if (err.name === "NotFoundError" || err.name === "DevicesNotFoundError") {
         errorType = "not-found";
-        message = "카메라를 찾을 수 없습니다. 테스트 이미지를 사용합니다.";
+        message = "카메라를 찾을 수 없습니다.";
       }
     } else if (typeof err === "string") {
       if (err.includes("Permission denied") || err.includes("Not allowed")) {
         errorType = "not-allowed";
-        message = "카메라 권한이 거부되었습니다. 테스트 이미지를 사용합니다.";
+        message = "카메라 권한이 거부되었습니다.";
       } else if (err.includes("Not found") || err.includes("No device")) {
         errorType = "not-found";
-        message = "카메라를 찾을 수 없습니다. 테스트 이미지를 사용합니다.";
+        message = "카메라를 찾을 수 없습니다.";
       }
     }
 
     setError({ type: errorType, message });
-    setUseTestImage(true);
-    setIsReady(true);
+    setIsReady(false);
   }, []);
 
   const requestPermission = useCallback(() => {
@@ -72,7 +68,6 @@ export function useCamera() {
     stopCamera();
     setIsRetrying(false);
     setError(null);
-    setUseTestImage(false);
     setPermissionRequested(false);
     setTimeout(() => setPermissionRequested(true), 100);
   }, [stopCamera]);
@@ -82,7 +77,6 @@ export function useCamera() {
     videoConstraints,
     isReady,
     error,
-    useTestImage,
     isRetrying,
     permissionRequested,
     stopCamera,
